+++ lib/search_object.rb
@@ -9,6 +9,12 @@ require 'search_object/plugin/kaminari'
 require 'search_object/plugin/sorting'
 
 module SearchObject
+  class MissingScopeError < RuntimeError
+    def initialize
+      super 'No scope provided. Scope can be defined on a class level or passed as an option.'
+    end
+  end
+
   def self.module(*plugins)
     return Base if plugins.empty?