+++ lib/search_object/base.rb
@@ -42,7 +42,7 @@ module SearchObject
     module ClassMethods
       # :api: private
       def build_internal_search(options)
-        scope  = options.fetch(:scope) { @scope.call }
+        scope  = options.fetch(:scope) { @scope && @scope.call } or raise MissingScopeError
         params = @defaults.merge(Helper.select_keys Helper.stringify_keys(options.fetch(:filters, {})), @actions.keys)
 
         Search.new scope, params, @actions