+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 1.0 (unreleased)
 
+* Raise `SearchObject::MissingScopeError` when no scope is provided
+
 * Replace position arguments with Hash of options (__backward incompatible__)
 
   ```diff